Add column field to `Token` struct
This is important because we use `Expr`s (which often contain a `Token`)
as the keys to the `locals` `HashMap` on the `Interpreter` struct. We
need each `Expr` to be unique. However, if there are, for example, two
occurrences of a particular variable on the same line, they are
indistinguishable. This means that they will map to the same local
variable and thus be resolved the same way. This causes a problem in
code like this:

```
if (x > 5) { return x; }
```

The second `x` is in a block which means that it shouldn't be resolved
to the same scope as the `x` in the if condition. This isn't a problem
if the then branch is on its own line. This can be fixed by adding a
`column` to disambiguate the two.