Define functions in the local environment
Since we aliased the initial environment to `globals`, we should change
this to just be the interpreter's current environment. When a function
is defined at the top level, it will be added to the global namespace
but if a function is defined in some other scope (e.g., inside another
function), it will now be resolved correctly.